refactor(edgeless): extract side effects of edgeless element toolbar (#8828)
This PR moves all `edgeless element toolbar` related side effects into a separate file.
