+++ packages/blocks/src/root-block/widgets/element-toolbar/align-button.ts
@@ -326,12 +326,6 @@ export class EdgelessAlignButton extends WithDisposable(LitElement) {
   accessor edgeless!: EdgelessRootBlockComponent;
 }
 
-declare global {
-  interface HTMLElementTagNameMap {
-    'edgeless-align-button': EdgelessAlignButton;
-  }
-}
-
 export function renderAlignButton(
   edgeless: EdgelessRootBlockComponent,
   elements: BlockSuite.EdgelessModel[]
