+++ packages/blocks/src/root-block/widgets/element-toolbar/change-image-button.ts
@@ -72,12 +72,6 @@ export class EdgelessChangeImageButton extends WithDisposable(LitElement) {
   accessor model!: ImageBlockModel;
 }
 
-declare global {
-  interface HTMLElementTagNameMap {
-    'edgeless-change-image-button': EdgelessChangeImageButton;
-  }
-}
-
 export function renderChangeImageButton(
   edgeless: EdgelessRootBlockComponent,
   images?: ImageBlockModel[]
