+++ packages/blocks/src/root-block/widgets/element-toolbar/release-from-group-button.ts
@@ -42,12 +42,6 @@ export class EdgelessReleaseFromGroupButton extends WithDisposable(LitElement) {
   accessor edgeless!: EdgelessRootBlockComponent;
 }
 
-declare global {
-  interface HTMLElementTagNameMap {
-    'edgeless-release-from-group-button': EdgelessReleaseFromGroupButton;
-  }
-}
-
 export function renderReleaseFromGroupButton(
   edgeless: EdgelessRootBlockComponent
 ) {
